+++ src/pfsys/mod.rs
@@ -29,13 +29,6 @@ use std::path::PathBuf;
 use std::time::Instant;
 use thiserror::Error as thisError;
 
-#[cfg(feature = "python-bindings")]
-use pyo3::prelude::*;
-#[cfg(feature = "python-bindings")]
-use pyo3::types::PyDict;
-#[cfg(feature = "python-bindings")]
-use pyo3::ToPyObject;
-
 #[derive(thisError, Debug)]
 /// Errors related to pfsys
 pub enum PfSysError {
@@ -50,35 +43,6 @@ pub enum TranscriptType {
     Blake,
     Poseidon,
     EVM,
-}
-
-/// Truncates nested vector due to omit junk floating point values in python
-#[cfg(feature = "python-bindings")]
-fn truncate_nested_vector(input: &Vec<Vec<f32>>) -> Vec<Vec<f32>> {
-    let mut input_mut = input.clone();
-    for inner_vec in input_mut.iter_mut() {
-        for value in inner_vec.iter_mut() {
-            // truncate 6 decimal places
-            *value = (*value * 10000000.0).trunc() / 10000000.0;
-        }
-    }
-    input_mut
-}
-
-#[cfg(feature = "python-bindings")]
-impl ToPyObject for GraphInput {
-    fn to_object(&self, py: Python) -> PyObject {
-        // Create a Python dictionary
-        let dict = PyDict::new(py);
-        let input_data_mut = &self.input_data;
-        let output_data_mut = &self.output_data;
-        dict.set_item("input_data", truncate_nested_vector(&input_data_mut))
-            .unwrap();
-        dict.set_item("output_data", truncate_nested_vector(&output_data_mut))
-            .unwrap();
-
-        dict.to_object(py)
-    }
 }
 
 /// Defines the proof generated by a model / circuit suitably for serialization/deserialization.
